API & Web/project/backend/requirements.txt
autopep8==1.6.0
distlib==0.3.4
Django==4.0.3
django-cors-headers==3.12.0
django-filter==21.1
djangorestframework==3.13.1
filelock==3.6.0
platformdirs==2.5.1
pycodestyle==2.8.0
six==1.16.0
toml==0.10.2
virtualenv==20.14.0
asgiref==3.5.0
certifi==2021.10.8
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
wincertstore==0.2
